Add files via upload
Added "handleAddTodoItem" function into "Todos.js" and brought it into "AddTodo.js" as "addTodoItem". A reference was put that grabbed the value submitted into a form, and that value was passed into "Todos.js" and into "handleAddTodoItem" via "task" which created a copy of state via prevState and rendered  a new todo object with the task that was typed in followed up by the rest of the todo items in state.